// dashboard/src/client/models/workflow.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { WorkflowEdge } from './workflow-edge';
// May contain unused imports in some cases
// @ts-ignore
import type { WorkflowMetadata } from './workflow-metadata';
// May contain unused imports in some cases
// @ts-ignore
import type { WorkflowNode } from './workflow-node';
// May contain unused imports in some cases
// @ts-ignore
import type { WorkflowSubgraph } from './workflow-subgraph';

export interface Workflow {
    /**
     * Workflow UUID
     */
    'id': string;
    /**
     * e.g. \"2025-10\"
     */
    'schemaVersion': string;
    'metadata': WorkflowMetadata;
    'nodes': Array<WorkflowNode>;
    'edges': Array<WorkflowEdge>;
    /**
     * Workflow-level tags.
     */
    'tags'?: Array<string>;
    /**
     * Base64-encoded preview of the workflow canvas.
     */
    'previewImage'?: string | null;
    /**
     * Reusable workflow fragments (localized snapshots) that container nodes can reference.
     */
    'subgraphs'?: Array<WorkflowSubgraph>;
}